+++ CHANGES.md
@@ -1,3 +1,9 @@
+## Changes in 1.4.0 (in development)
+
+- Added opening parameters to the attributes of returned datasets.
+- Improved attribute handling for Sentinel-3 data access.
+
+
 ## Changes in 1.3.3
 
 - Set the correct time-axis encoding for ARDC requests, mitigating Zarr serialization